+++ Sensor/Adapter.SQLLit/Models/SensorState.cs
@@ -7,10 +7,10 @@ public class SensorState
 {
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public int Id { get; set; }
-    public int Value { get; set; }
+    public sbyte Value { get; set; }
     public DateTime DateTime { get; set; }
     
-    public SensorState(int value,DateTime dateTime)
+    public SensorState(sbyte value,DateTime dateTime)
     {
         Value = value;
         DateTime = dateTime;
